refactor(tui): split collector into submodules
No functional changes.

- Convert `src/collector/tui.rs` into a directory module with focused submodules (mod/terminal/state/render/input).
- Move the tracing log overlay into feature-gated `tui_log`.
- Gate feature-dependent examples via Cargo `required-features` so `--no-default-features` builds cleanly.
