--- src/collector/tui.rs
@@ -1,760 +1,0 @@
-//! Terminal user interface (TUI) collector for real-time benchmark monitoring.
-//!
-//! This module provides [`TuiCollector`], an interactive terminal-based collector
-//! that displays real-time benchmark statistics with a rich visual interface.
-//!
-//! # Features
-//!
-//! - Real-time statistics display (iteration rate, throughput, latency)
-//! - Rolling window statistics at multiple time scales (1s, 10s, 1min, 10min)
-//! - Latency histogram with percentiles
-//! - Iteration histogram showing throughput over time
-//! - Status distribution breakdown
-//! - Progress bar with duration/iteration tracking
-//! - Pause/resume support
-//! - Optional log viewer (with `tracing` feature)
-//!
-//! # Keyboard Controls
-//!
-//! - `+`/`-`: Zoom time window in/out
-//! - `a`: Auto-select time window based on elapsed time
-//! - `p`: Pause/resume the benchmark
-//! - `l`: Toggle log viewer (requires `tracing` feature)
-//! - `q` or `Ctrl+C`: Quit the benchmark
-
-use anyhow::Result;
-use async_trait::async_trait;
-use crossterm::{
-    ExecutableCommand, cursor,
-    event::{Event, KeyCode, KeyEvent, KeyModifiers},
-    terminal,
-};
-use itertools::Itertools;
-use nonzero_ext::nonzero;
-use ratatui::{
-    CompletedFrame, Frame,
-    backend::CrosstermBackend,
-    layout::{Constraint, Direction, Layout, Margin, Rect},
-    style::{Color, Style, Stylize},
-    text::Line,
-    widgets::{BarChart, Block, Borders, Clear, Gauge, Padding, Paragraph},
-};
-use std::{collections::HashMap, fmt, io, num::NonZeroU8, time::Duration};
-use tokio::{
-    sync::{mpsc, watch},
-    time::MissedTickBehavior,
-};
-use tokio_util::sync::CancellationToken;
-
-use crate::{
-    collector::ReportCollector,
-    duration::DurationExt,
-    histogram::{LatencyHistogram, PERCENTAGES},
-    report::{BenchReport, IterReport},
-    runner::BenchOpts,
-    stats::{Counter, IterStats, RotateDiffWindowGroup, RotateWindowGroup},
-    status::{Status, StatusKind},
-    util::{IntoAdjustedByte, TryIntoAdjustedByte},
-};
-
-const SECOND: Duration = Duration::from_secs(1);
-
-/// A report collector with real-time terminal user interface (TUI) support.
-///
-/// This collector displays a live dashboard showing benchmark progress,
-/// statistics, and histograms. It supports interactive controls for
-/// pausing, zooming time windows, and viewing logs.
-///
-/// The TUI uses [ratatui](https://ratatui.rs) for rendering and updates
-/// at the configured frame rate.
-///
-/// # Display Sections
-///
-/// - **Stats for last N**: Rolling window statistics (configurable via `+`/`-`)
-/// - **Stats overall**: Cumulative statistics since benchmark start
-/// - **Status distribution**: Breakdown of response statuses
-/// - **Iteration histogram**: Bar chart of iterations per time bucket
-/// - **Latency histogram**: Distribution of response latencies with percentiles
-/// - **Progress**: Progress bar showing completion status
-pub struct TuiCollector {
-    /// The benchmark options (duration, iterations, concurrency, etc.).
-    pub(crate) bench_opts: BenchOpts,
-    /// Refresh rate in frames per second (fps).
-    pub(crate) fps: NonZeroU8,
-    /// Channel receiver for iteration reports from workers.
-    pub(crate) res_rx: mpsc::UnboundedReceiver<Result<IterReport>>,
-    /// Watch channel sender for pause/resume control.
-    pub(crate) pause: watch::Sender<bool>,
-    /// Cancellation token for graceful shutdown.
-    pub(crate) cancel: CancellationToken,
-    /// Whether to exit automatically when the benchmark finishes.
-    pub(crate) auto_quit: bool,
-
-    /// Internal TUI state (time window selection, log state, etc.).
-    state: TuiCollectorState,
-}
-
-struct TuiCollectorState {
-    tm_win: TimeWindow,
-    finished: bool,
-    #[cfg(feature = "tracing")]
-    log: tui_log::LogState,
-}
-
-impl TuiCollector {
-    /// Create a new TUI report collector.
-    pub fn new(
-        bench_opts: BenchOpts,
-        fps: NonZeroU8,
-        res_rx: mpsc::UnboundedReceiver<Result<IterReport>>,
-        pause: watch::Sender<bool>,
-        cancel: CancellationToken,
-        auto_quit: bool,
-    ) -> Result<Self> {
-        let state = TuiCollectorState {
-            tm_win: TimeWindow::Second,
-            finished: false,
-            #[cfg(feature = "tracing")]
-            log: tui_log::LogState::from_env()?,
-        };
-        Ok(Self { bench_opts, fps, res_rx, pause, cancel, auto_quit, state })
-    }
-}
-
-struct Terminal {
-    terminal: ratatui::Terminal<CrosstermBackend<io::Stdout>>,
-}
-
-impl Terminal {
-    fn new() -> Result<Self> {
-        crossterm::terminal::enable_raw_mode()?;
-        io::stdout().execute(crossterm::cursor::Hide)?;
-        io::stdout().execute(crossterm::terminal::EnterAlternateScreen)?;
-
-        Ok(Self {
-            terminal: ratatui::Terminal::new(CrosstermBackend::new(io::stdout()))?,
-        })
-    }
-
-    fn draw<F>(&mut self, f: F) -> io::Result<CompletedFrame<'_>>
-    where
-        F: FnOnce(&mut Frame),
-    {
-        self.terminal.draw(f)
-    }
-}
-
-impl Drop for Terminal {
-    fn drop(&mut self) {
-        let _ = std::io::stdout().execute(terminal::LeaveAlternateScreen);
-        let _ = std::io::stdout().execute(cursor::Show);
-        let _ = crossterm::terminal::disable_raw_mode();
-    }
-}
-
-#[async_trait]
-impl ReportCollector for TuiCollector {
-    async fn run(&mut self) -> Result<BenchReport> {
-        let mut hist = LatencyHistogram::new();
-        let mut stats = IterStats::new();
-        let mut status_dist = HashMap::new();
-        let mut error_dist = HashMap::new();
-
-        self.collect(&mut hist, &mut stats, &mut status_dist, &mut error_dist)
-            .await?;
-
-        let elapsed = self.bench_opts.clock.elapsed();
-        let concurrency = self.bench_opts.concurrency;
-        Ok(BenchReport { concurrency, hist, stats, status_dist, error_dist, elapsed })
-    }
-}
-
-impl TuiCollector {
-    async fn collect(
-        &mut self,
-        hist: &mut LatencyHistogram,
-        stats: &mut IterStats,
-        status_dist: &mut HashMap<Status, u64>,
-        error_dist: &mut HashMap<String, u64>,
-    ) -> Result<()> {
-        let clock = self.bench_opts.clock.clone();
-        let mut terminal = Terminal::new()?;
-
-        let mut latest_iters = RotateWindowGroup::new(nonzero!(60usize));
-        let mut latest_iters_ticker = clock.ticker(SECOND);
-
-        let mut latest_stats = RotateDiffWindowGroup::new(self.fps.into());
-        let mut latest_stats_ticker = clock.ticker(SECOND / self.fps.get() as u32);
-
-        let mut ui_ticker = tokio::time::interval(SECOND / self.fps.get() as u32);
-        ui_ticker.set_missed_tick_behavior(MissedTickBehavior::Burst);
-
-        loop {
-            if self.state.finished {
-                if self.auto_quit {
-                    return Ok(());
-                }
-                ui_ticker.tick().await;
-            } else {
-                loop {
-                    tokio::select! {
-                        biased;
-                        _ = ui_ticker.tick() => break,
-                        _ = latest_stats_ticker.tick() => {
-                            latest_stats.rotate(stats);
-                            continue;
-                        }
-                        _ = latest_iters_ticker.tick() => {
-                            latest_iters.rotate();
-                            continue;
-                        }
-                        r = self.res_rx.recv() => match r {
-                            Some(Ok(report)) => {
-                                *status_dist.entry(report.status).or_default() += 1;
-                                hist.record(report.duration)?;
-                                latest_iters.push(&report);
-                                *stats += &report;
-                            }
-                            Some(Err(e)) => *error_dist.entry(e.to_string()).or_default() += 1,
-                            None => {
-                                clock.pause();
-                                self.state.finished = true;
-                                break;
-                            }
-                        }
-                    };
-                }
-            }
-
-            let elapsed = clock.elapsed();
-            if self.handle_event(elapsed).await? {
-                return Ok(());
-            }
-
-            terminal.draw(|f| {
-                let progress_height = 3;
-                let stats_height = 5;
-                let error_dist_height = match error_dist.len() {
-                    0 => 0,
-                    len => len.min(5) as u16 + 2,
-                };
-                let hist_height_filler = 40;
-                let tips_height = 1;
-                let rows = Layout::default()
-                    .direction(Direction::Vertical)
-                    .constraints([
-                        Constraint::Length(stats_height),
-                        Constraint::Length(error_dist_height),
-                        Constraint::Fill(hist_height_filler),
-                        Constraint::Length(progress_height),
-                        Constraint::Length(tips_height),
-                    ])
-                    .split(f.area());
-
-                let mid = Layout::default()
-                    .direction(Direction::Horizontal)
-                    .constraints([Constraint::Fill(1), Constraint::Fill(1), Constraint::Fill(1)])
-                    .split(rows[0]);
-
-                let bot = Layout::default()
-                    .direction(Direction::Horizontal)
-                    .constraints([Constraint::Percentage(50), Constraint::Percentage(50)].as_ref())
-                    .split(rows[2]);
-
-                let paused = *self.pause.borrow();
-                let finished = self.state.finished;
-                render_process_gauge(f, rows[3], &stats.counter, elapsed, &self.bench_opts, paused, finished);
-                render_stats_overall(f, mid[1], &stats.counter, elapsed);
-                render_stats_timewin(f, mid[0], &latest_stats, self.state.tm_win);
-                render_status_dist(f, mid[2], status_dist);
-                render_error_dist(f, rows[1], error_dist);
-                render_iter_hist(f, bot[0], &latest_iters, self.state.tm_win);
-                render_latency_hist(f, bot[1], hist, 7);
-                render_tips(f, rows[4]);
-
-                #[cfg(feature = "tracing")]
-                tui_log::render_logs(f, &self.state.log);
-            })?;
-        }
-    }
-
-    /// Handle the user input events. Returns `true` if the collector should quit.
-    async fn handle_event(&mut self, elapsed: Duration) -> Result<bool> {
-        let clock = &mut self.bench_opts.clock;
-        while crossterm::event::poll(Duration::from_secs(0))? {
-            use KeyCode::*;
-            if let Event::Key(KeyEvent { code, modifiers, .. }) = crossterm::event::read()? {
-                match (code, modifiers) {
-                    (Char('+'), _) => {
-                        self.state.tm_win = self.state.tm_win.prev();
-                    }
-                    (Char('-'), _) => {
-                        self.state.tm_win = self.state.tm_win.next();
-                    }
-                    (Char('a'), _) => {
-                        self.state.tm_win = *TimeWindow::variants()
-                            .iter()
-                            .rfind(|&&ts| elapsed > Duration::from(ts))
-                            .unwrap_or(&TimeWindow::Second)
-                    }
-                    (Char('q'), _) | (Char('c'), KeyModifiers::CONTROL) => {
-                        self.cancel.cancel();
-                        return Ok(true);
-                    }
-                    (Char('p') | Pause, _) if !self.state.finished => {
-                        let pause = !*self.pause.borrow();
-                        if pause {
-                            clock.pause();
-                        } else {
-                            clock.resume();
-                        }
-                        self.pause.send_replace(pause);
-                    }
-                    #[cfg(feature = "tracing")]
-                    (Char('l'), _) => self.state.log.display = !self.state.log.display,
-                    #[cfg(feature = "tracing")]
-                    (code, _) if self.state.log.display => {
-                        use tui_logger::TuiWidgetEvent::*;
-                        let txn = |e| self.state.log.inner.transition(e);
-                        match code {
-                            Char(' ') => txn(HideKey),
-                            PageDown | Char('f') => txn(NextPageKey),
-                            PageUp | Char('b') => txn(PrevPageKey),
-                            Up => txn(UpKey),
-                            Down => txn(DownKey),
-                            Left => txn(LeftKey),
-                            Right => txn(RightKey),
-                            Enter => txn(FocusKey),
-                            Esc => txn(EscapeKey),
-                            _ => (),
-                        }
-                    }
-                    _ => (),
-                }
-            }
-        }
-        Ok(false)
-    }
-}
-
-fn render_stats_timewin(frame: &mut Frame, area: Rect, stats: &RotateDiffWindowGroup, tw: TimeWindow) {
-    let (stats, duration) = match tw {
-        TimeWindow::Second => stats.stats_last_sec(),
-        TimeWindow::TenSec => stats.stats_last_10sec(),
-        TimeWindow::Minute => stats.stats_last_min(),
-        TimeWindow::TenMin => stats.stats_last_10min(),
-    };
-
-    render_stats(
-        frame,
-        area,
-        Line::from(vec!["Stats for ".into(), format!("last {}", tw).yellow().bold()]),
-        &stats.counter,
-        duration,
-    );
-}
-
-fn render_stats_overall(frame: &mut Frame, area: Rect, counter: &Counter, elapsed: Duration) {
-    render_stats(frame, area, "Stats overall".into(), counter, elapsed);
-}
-
-fn render_stats(frame: &mut Frame, area: Rect, title: Line<'_>, counter: &Counter, elapsed: Duration) {
-    let block = Block::new().title(title).borders(Borders::ALL);
-
-    let [lhs, rhs] =
-        Layout::horizontal([Constraint::Percentage(50), Constraint::Percentage(50)]).areas(block.inner(area));
-
-    let stats_counter = render_stats_counter(counter);
-    let stats_rate = render_stats_rate(counter, elapsed);
-
-    frame.render_widget(stats_counter, lhs);
-    frame.render_widget(stats_rate, rhs);
-    frame.render_widget(block, area);
-}
-
-fn render_stats_counter(counter: &Counter) -> Paragraph<'static> {
-    let lines = vec![
-        Line::from(vec!["Items: ".into(), counter.items.to_string().green()]),
-        Line::from(vec!["Iters: ".into(), counter.iters.to_string().green()]),
-        Line::from(vec![
-            "Bytes: ".into(),
-            format!("{:.2}", counter.bytes.adjusted()).green(),
-        ]),
-    ];
-    Paragraph::new(lines).block(Block::new().borders(Borders::NONE))
-}
-
-fn render_stats_rate(counter: &Counter, elapsed: Duration) -> Paragraph<'static> {
-    let secs = elapsed.as_secs_f64();
-    let lines = vec![
-        Line::from(format!("{:.2} iters/s", counter.iters as f64 / secs).green()),
-        Line::from(format!("{:.2} items/s", counter.items as f64 / secs).green()),
-        Line::from(
-            format!(
-                "{}/s",
-                match (counter.bytes as f64 / secs).adjusted() {
-                    Ok(bps) => format!("{:.2}", bps),
-                    Err(_) => "NaN B".to_string(),
-                }
-            )
-            .green(),
-        ),
-    ];
-    Paragraph::new(lines).block(Block::new().borders(Borders::NONE))
-}
-
-fn render_process_gauge(
-    frame: &mut Frame,
-    area: Rect,
-    counter: &Counter,
-    elapsed: Duration,
-    opts: &BenchOpts,
-    paused: bool,
-    finished: bool,
-) {
-    let rounded = |duration: Duration| humantime::Duration::from(Duration::from_secs(duration.as_secs_f64() as u64));
-    let time_progress = |duration: &Duration| {
-        (
-            (elapsed.as_secs_f64() / duration.as_secs_f64()).clamp(0.0, 1.0),
-            format!("{} / {}", rounded(elapsed), rounded(*duration)),
-        )
-    };
-    let iter_progress = |iters: &u64| {
-        (
-            (counter.iters as f64 / *iters as f64).clamp(0.0, 1.0),
-            format!("{} / {}", counter.iters, iters),
-        )
-    };
-
-    let (progress, mut label) = match opts {
-        BenchOpts { duration: None, iterations: None, .. } => (0.0, "INFINITE".to_string()),
-        BenchOpts { duration: Some(duration), iterations: None, .. } => time_progress(duration),
-        BenchOpts { duration: None, iterations: Some(iters), .. } => iter_progress(iters),
-        BenchOpts { duration: Some(duration), iterations: Some(iters), .. } => {
-            let iter_ratio = counter.iters as f64 / *iters as f64;
-            let time_ratio = elapsed.as_secs_f64() / duration.as_secs_f64();
-            if iter_ratio > time_ratio {
-                iter_progress(iters)
-            } else {
-                time_progress(duration)
-            }
-        }
-    };
-
-    let style = match (finished, paused) {
-        (true, _) => {
-            label.push_str(" (FINISHED)");
-            Style::new().fg(Color::Yellow)
-        }
-        (_, true) => {
-            label.push_str(" (PAUSED)");
-            Style::new().fg(Color::Yellow)
-        }
-        (false, false) => Style::new().fg(Color::Cyan),
-    };
-
-    let guage = Gauge::default()
-        .block(Block::new().title("Progress").borders(Borders::ALL))
-        .gauge_style(style)
-        .label(label)
-        .ratio(progress);
-    frame.render_widget(guage, area);
-}
-
-fn render_status_dist(frame: &mut Frame, area: Rect, status_dist: &HashMap<Status, u64>) {
-    let dist = status_dist
-        .iter()
-        .sorted_by_key(|&(_, cnt)| std::cmp::Reverse(cnt))
-        .map(|(status, cnt)| {
-            let s = format!("{} {} iters", status, cnt);
-            let s = match status.kind() {
-                StatusKind::Success => s.green(),
-                StatusKind::ClientError => s.yellow(),
-                StatusKind::ServerError => s.red(),
-                StatusKind::Error => s.magenta(),
-            };
-            Line::from(s)
-        })
-        .collect_vec();
-    let p = Paragraph::new(dist).block(Block::new().title("Status distribution").borders(Borders::ALL));
-    frame.render_widget(p, area);
-}
-
-fn render_error_dist(frame: &mut Frame, area: Rect, error_dist: &HashMap<String, u64>) {
-    if error_dist.is_empty() {
-        return;
-    }
-
-    let dist = error_dist
-        .iter()
-        .sorted_by_key(|&(_, cnt)| std::cmp::Reverse(cnt))
-        .map(|(err, cnt)| Line::from(format!("[{cnt}] {err}")))
-        .collect_vec();
-    let p = Paragraph::new(dist).block(Block::new().title("Error distribution").borders(Borders::ALL));
-    frame.render_widget(p, area);
-}
-
-fn render_iter_hist(frame: &mut Frame, area: Rect, rwg: &RotateWindowGroup, tw: TimeWindow) {
-    let win = match tw {
-        TimeWindow::Second => &rwg.stats_by_sec,
-        TimeWindow::TenSec => &rwg.stats_by_10sec,
-        TimeWindow::Minute => &rwg.stats_by_min,
-        TimeWindow::TenMin => &rwg.stats_by_10min,
-    };
-    let cols = win.iter().map(|w| w.counter.iters.to_string().len()).max().unwrap_or(0);
-    let data = win
-        .iter()
-        .enumerate()
-        .map(|(i, n)| {
-            let mut s = tw.format(i);
-            if cols > s.len() {
-                for _ in 0..cols - s.len() {
-                    s.push(' ');
-                }
-            }
-            (s, n.counter.iters)
-        })
-        .collect_vec();
-
-    let bar_num_iter_str = data.iter().map(|(a, b)| (a.as_str(), *b)).collect_vec();
-    let bar_width = data
-        .iter()
-        .map(|(s, _)| s.chars().count())
-        .max()
-        .map(|w| w + 2)
-        .unwrap_or(1) as u16;
-    let chart = BarChart::default()
-        .block(Block::new().title("Iteration histogram").borders(Borders::ALL))
-        .data(bar_num_iter_str.as_slice())
-        .bar_style(Style::default().fg(Color::Green))
-        .label_style(Style::default().fg(Color::Cyan))
-        .bar_width(bar_width);
-    frame.render_widget(chart, area);
-}
-
-fn render_latency_hist(frame: &mut Frame, area: Rect, hist: &LatencyHistogram, histo_width: usize) {
-    // time unit for the histogram
-    let u = hist.median().appropriate_unit();
-
-    let quantiles = hist
-        .quantiles()
-        .map(|(d, n)| (d.as_f64(u).to_string(), n))
-        .collect_vec();
-
-    let data = quantiles.iter().map(|(d, n)| (d.as_str(), *n)).collect_vec();
-    let chart = BarChart::default()
-        .block(
-            Block::new()
-                .title(Line::from(vec![
-                    "Latency histogram (".into(),
-                    u.to_string().yellow().bold(),
-                    ")".into(),
-                ]))
-                .borders(Borders::ALL),
-        )
-        .data(&data)
-        .bar_style(Style::default().fg(Color::Green))
-        .label_style(Style::default().fg(Color::Cyan))
-        .bar_width(histo_width as u16);
-    frame.render_widget(chart, area);
-
-    if hist.is_empty() {
-        return;
-    }
-
-    let area = area.inner(Margin::new(1, 1));
-    let area = Layout::horizontal([Constraint::Percentage(50), Constraint::Percentage(50)]).split(area);
-    let area = Layout::vertical([Constraint::Percentage(50), Constraint::Percentage(50)]).split(area[1]);
-    let area = area[0];
-
-    // max width of the formatted duration
-    let w = format!("{:.2}", hist.max().as_f64(u)).len();
-    #[rustfmt::skip]
-    let mut content = vec![
-        Line::from(vec!["Avg: ".blue(),      format!("{: >w$.2}", hist.mean().as_f64(u)).green()]),
-        Line::from(vec!["Min: ".cyan(),      format!("{: >w$.2}", hist.min().as_f64(u)).green()]),
-        Line::from(vec!["Med: ".yellow(),    format!("{: >w$.2}", hist.median().as_f64(u)).green()]),
-        Line::from(vec!["Max: ".red(),       format!("{: >w$.2}", hist.max().as_f64(u)).green()]),
-        Line::from(vec!["Stdev: ".magenta(), format!("{: >w$.2}", hist.stdev().as_f64(u)).green()]),
-    ];
-    content.push(Line::default());
-
-    content.extend(hist.percentiles(PERCENTAGES).map(|(p, d)| {
-        Line::from(vec![
-            format!("P{:.2}%: ", p).cyan(),
-            format!("{: >w$.2}", d.as_f64(u)).green(),
-        ])
-    }));
-    let width = content.iter().map(|s| s.width()).max().unwrap_or(0) + 2;
-    if width > area.width as usize {
-        return;
-    }
-    let area = Rect {
-        x: area.x + area.width - width as u16,
-        y: area.y,
-        width: width as u16,
-        height: content.len() as u16,
-    };
-    let block = Block::default().padding(Padding::right(2)).borders(Borders::NONE);
-    let paragraph = Paragraph::new(content).block(block).right_aligned();
-
-    frame.render_widget(Clear, area); //clears out the background
-    frame.render_widget(paragraph, area);
-}
-
-fn gen_tips<'a>(tips: impl IntoIterator<Item = (&'a str, &'a str)>) -> Line<'a> {
-    #[allow(unstable_name_collisions)]
-    tips.into_iter()
-        .map(|(key, tip)| vec![key.bold().yellow().italic(), ": ".into(), tip.italic()])
-        .intersperse(vec![", ".into()])
-        .flatten()
-        .collect_vec()
-        .into()
-}
-
-fn render_tips(frame: &mut Frame, area: Rect) {
-    let tips = gen_tips([
-        ("+/-", "Zoom in/out"),
-        ("a", "Auto time window"),
-        #[cfg(feature = "tracing")]
-        ("l", "Logs window"),
-        ("p", "Pause"),
-        ("q", "Quit"),
-    ])
-    .right_aligned();
-    frame.render_widget(tips, area.inner(Margin::new(1, 0)));
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub enum TimeWindow {
-    Second = 1,
-    TenSec = 10,
-    Minute = 60,
-    TenMin = 600,
-}
-
-impl TimeWindow {
-    fn variants() -> &'static [TimeWindow] {
-        use TimeWindow::*;
-        &[Second, TenSec, Minute, TenMin]
-    }
-}
-
-impl fmt::Display for TimeWindow {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", humantime::format_duration(Duration::from(*self)))
-    }
-}
-
-impl From<TimeWindow> for Duration {
-    fn from(tw: TimeWindow) -> Self {
-        Duration::from_secs(tw as u64)
-    }
-}
-
-impl TimeWindow {
-    pub fn format(&self, n: usize) -> String {
-        match self {
-            TimeWindow::Second => format!("{}s", n),
-            TimeWindow::TenSec => format!("{}s", 10 * n),
-            TimeWindow::Minute => format!("{}m", n),
-            TimeWindow::TenMin => format!("{}m", 10 * n),
-        }
-    }
-
-    pub fn next(&self) -> Self {
-        match self {
-            TimeWindow::Second => TimeWindow::TenSec,
-            TimeWindow::TenSec => TimeWindow::Minute,
-            TimeWindow::Minute => TimeWindow::TenMin,
-            TimeWindow::TenMin => TimeWindow::TenMin,
-        }
-    }
-
-    pub fn prev(&self) -> Self {
-        match self {
-            TimeWindow::Second => TimeWindow::Second,
-            TimeWindow::TenSec => TimeWindow::Second,
-            TimeWindow::Minute => TimeWindow::TenSec,
-            TimeWindow::TenMin => TimeWindow::Minute,
-        }
-    }
-}
-
-#[cfg(feature = "tracing")]
-mod tui_log {
-    use super::*;
-
-    use log::LevelFilter;
-    use tui_logger::{TuiLoggerLevelOutput, TuiLoggerSmartWidget, TuiWidgetState};
-
-    pub(crate) struct LogState {
-        pub(crate) inner: TuiWidgetState,
-        pub(crate) display: bool,
-    }
-
-    impl LogState {
-        pub(crate) fn from_env() -> Result<Self> {
-            tui_logger::set_default_level(LevelFilter::Trace);
-            let state = TuiWidgetState::new().set_default_display_level(LevelFilter::Info);
-            Ok(Self { inner: state, display: false })
-        }
-    }
-
-    pub(crate) fn render_logs(frame: &mut Frame, state: &LogState) {
-        if !state.display {
-            return;
-        }
-
-        let log_widget = TuiLoggerSmartWidget::default()
-            .style_error(Style::default().fg(Color::Red))
-            .style_debug(Style::default().fg(Color::Green))
-            .style_warn(Style::default().fg(Color::Yellow))
-            .style_trace(Style::default().fg(Color::Magenta))
-            .style_info(Style::default().fg(Color::Cyan))
-            .border_type(ratatui::widgets::BorderType::Rounded)
-            .output_separator('|')
-            .output_level(Some(TuiLoggerLevelOutput::Abbreviated))
-            .output_target(true)
-            .output_file(true)
-            .output_line(true)
-            .title_log("Logs")
-            .title_target("Selector")
-            .state(&state.inner);
-
-        let area = centered_rect(80, 80, frame.area());
-        let rows = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([Constraint::Percentage(100), Constraint::Min(1)])
-            .split(area.inner(Margin::new(1, 1)));
-        let tips = gen_tips([
-            ("Enter", "Focus target"),
-            ("↑/↓", "Select target"),
-            ("←/→", "Display level"),
-            ("f/b", "Scroll"),
-            ("Esc", "Cancel scroll"),
-            ("Space", "Hide selector"),
-        ])
-        .right_aligned();
-
-        frame.render_widget(Clear, area);
-        frame.render_widget(log_widget, rows[0]);
-        frame.render_widget(tips, rows[1].inner(Margin::new(1, 0)));
-    }
-
-    pub(crate) fn centered_rect(percent_x: u16, percent_y: u16, r: Rect) -> Rect {
-        let popup_layout = Layout::vertical([
-            Constraint::Percentage((100 - percent_y) / 2),
-            Constraint::Percentage(percent_y),
-            Constraint::Percentage((100 - percent_y) / 2),
-        ])
-        .split(r);
-
-        Layout::horizontal([
-            Constraint::Percentage((100 - percent_x) / 2),
-            Constraint::Percentage(percent_x),
-            Constraint::Percentage((100 - percent_x) / 2),
-        ])
-        .split(popup_layout[1])[1]
-    }
-}
